Saveable Components and GameObjects
Unique identifier u sed to make the data always associated with entity

TODO: When loading  a  state for player, the camera goes out of control for a moment. This needs to be fixed, as the player is able to see off of the map. This is not good for gameplay experience and immersion.